refactor(lab): extract offload fallback helpers
